Add new controller logic for the Get Involved Page
The core logic of the controller is a migration of a set of function
calls in whitehall/app/controllers/home_controller.rb for populating the
Get Involved page with relevant data.

In this case the function calls are replaced with calls to the
publishing-api and respective filtering of those results to give us what
we need in Government Frontend.