Add helper and presenter for Get Involved migration
The helper in this case handles a selection of time to string filtering/
labelling and a single routing function to allow us to link to the
search page for a given set of consultations (feature match to Whitehall
original version).

The presenter is empty and is included for consistency with other
content.